fix authentication done immediately after signup
On intial signup, looking up credentials by providerId can result in no records being found since the query was not performed by the key. To get around this the account key is included in the /auth request immediately after signup, allowing the credentials to be looked up by the account ancestor. Once the account is found the credentials are then validated to belong to the account.